[eclipse/xtext#1224] Detect local Jenkins environment
Build steps defined in Jenkinsfile pass the built-in environment variable 'JENKINS_URL' to the Gradle/Maven executions. This is evaluated in the build scripts for upstream repository URLs. On Xtext JIPP this will use upstream repos from JIPP. In local builds outside of Jenkins the property defaults to Typefox CI like before.

Signed-off-by: Karsten Thoms <karsten.thoms@itemis.de>